+++ packages/editor/editor-core/src/__tests__/visual-regression/common/breakout.ts
@@ -2,7 +2,7 @@ import { initFullPageEditorWithAdf, Device, snapshot } from '../_utils';
 import { KeyboardKeys } from '../../__helpers/page-objects/_keyboard';
 import * as adf from './__fixtures__/columns.adf.json';
 
-describe('Snapshot Test: Popup', () => {
+describe.skip('Snapshot Test: Popup', () => {
   it('placement: start,end is correct on scroll', async () => {
     // @ts-ignore
     const page = global.page;