+++ packages/elements/emoji/src/__tests__/unit/ssr/server-side.ts
@@ -1,28 +1,11 @@
 /**
  * @jest-environment node
  */
-
 import { SSRHelper } from '@atlaskit/elements-test-helpers';
-import { whiteList } from './_ssr-config';
-import 'whatwg-fetch';
+import { blackList } from './_ssr-config';
 
 describe('server side rendering', () => {
-  const ssrHelper = new SSRHelper(whiteList);
-  let _window: any;
-
-  beforeAll(() => {
-    if (global) {
-      _window = (global as any).window;
-      // note: emoji resource uses localStorage if available so gotta mock window to avoid errors there
-      (global as any).window = {};
-    }
-  });
-
-  afterAll(() => {
-    if (global) {
-      (global as any).window = _window;
-    }
-  });
+  const ssrHelper = new SSRHelper({ blackList });
 
   test('emoji server side rendering', async () => {
     await ssrHelper.renderSSRAndAssert('emoji');