+++ packages/elements/status/src/__tests__/unit/ssr/hydrate.ts
@@ -1,15 +1,7 @@
-import { SSRHelper } from '@atlaskit/elements-test-helpers';
+import { SSRHydrationHelper } from '@atlaskit/elements-test-helpers';
 
-describe('SSR', () => {
-  const ssrHelper = new SSRHelper();
-
-  beforeAll(() => {
-    ssrHelper.beforeHydration();
-  });
-
-  afterAll(() => {
-    ssrHelper.afterHydration();
-  });
+describe('server side rendering - hydrate', () => {
+  const ssrHelper = new SSRHydrationHelper({});
 
   test('should ssr then hydrate tag correctly', async () => {
     await ssrHelper.hydrateSSRAndAssert('status');