+++ packages/core/navigation-next/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @atlaskit/navigation-next
 
+## 4.4.1
+- [patch] [e71cc26e15](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/e71cc26e15):
+
+  - Change padding-top of SectionHeading to margin-top
+
+## 4.4.0
+- [minor] [8ebfb7165d](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/8ebfb7165d):
+
+  - New props: topOffset - The top offset value to be used in navigation
+
+## 4.3.1
+- [patch] [5cf4e22048](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/5cf4e22048):
+
+  - Fixes the issue where the switcher stayed open while navigation is being resized.
+  - Fixes the issue where collapse/expanding the navigation caused the Switcher to reset to it's default size.
+
+  [Jira ticket NAV-237](https://product-fabric.atlassian.net/browse/NAV-237)
+
 ## 4.3.0
 - [minor] [f54655ac1a](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/f54655ac1a):